move SBMLNonspatialSimResults from vcell-cli to vcell-core
