+++ vcell-cli/src/main/java/org/vcell/cli/run/SolverHandler.java
@@ -6,8 +6,6 @@ import cbit.vcell.biomodel.BioModel;
 import cbit.vcell.field.FieldDataIdentifierSpec;
 import cbit.vcell.mapping.MathSymbolMapping;
 import cbit.vcell.math.FunctionColumnDescription;
-import cbit.vcell.math.MathDescription;
-import cbit.vcell.math.SourceSymbolMapping;
 import cbit.vcell.messaging.server.SimulationTask;
 import cbit.vcell.parser.Expression;
 import cbit.vcell.parser.ExpressionException;
@@ -23,7 +21,6 @@ import cbit.vcell.solver.stoch.GibsonSolver;
 import cbit.vcell.solver.stoch.HybridSolver;
 import cbit.vcell.solvers.AbstractCompiledSolver;
 import cbit.vcell.xml.ExternalDocInfo;
-import cbit.vcell.xml.XmlHelper;
 
 import org.jlibsedml.AbstractTask;
 import org.jlibsedml.DataGenerator;
@@ -44,6 +41,7 @@ import org.jmathml.ASTNode;
 import org.vcell.cli.CLIRecordable;
 import org.vcell.sbml.vcell.SBMLImportException;
 import org.vcell.sbml.vcell.SBMLImporter;
+import org.vcell.sbml.vcell.SBMLNonspatialSimResults;
 import org.vcell.sbml.vcell.SBMLSymbolMapping;
 import org.vcell.sedml.SEDMLImporter;
 import org.vcell.util.ISize;
